requirements.txt
# Obsidian Auto-Linker Dependencies
# Core dependencies for AI-powered Obsidian vault processing
=======
# Enhanced Obsidian Auto-Linker Dependenciesas
# Core dependencies for AI-powered Obsidian vault processing
 main

# Core dependencies
requests>=2.25.0
pyyaml>=6.0
tqdm>=4.65.0
psutil>=5.8.0
rich>=13.0.0

# HTTP requests for Ollama API
requests>=2.32.0

# Progress bars and terminal UI
tqdm>=4.67.0

# System monitoring and process management
psutil>=7.1.0

# File system operations and path handling
pathlib2>=2.3.7; python_version < "3.4"
typing-extensions>=4.0.0; python_version < "3.8"

# Testing dependencies
pytest>=7.0.0
pytest-cov>=4.0.0

# Configuration validation
pydantic>=2.0.0

# JSON handling (built-in, but ensuring compatibility)
# json - built-in module

# Regular expressions (built-in)
# re - built-in module

# Date and time handling (built-in)
# datetime - built-in module

# Threading and concurrency (built-in)
# threading - built-in module
# concurrent.futures - built-in module

# File operations (built-in)
# os - built-in module
# shutil - built-in module
# hashlib - built-in module

# Time operations (built-in)
# time - built-in module

# Type hints (built-in)
# typing - built-in module

# Optional: Enhanced terminal output
colorama>=0.4.6

# Optional: Better YAML handling
ruamel.yaml>=0.17.0

 Optional: Enhanced file watching
watchdog>=3.0.0

 Optional: Better progress reporting
rich>=13.0.0

 Development dependencies (optional)
 pytest>=7.0.0
 black>=23.0.0
 flake8>=6.0.0
